初步完成报告并融合stride,添加一些debug报告
初步完成报告，并实现stride比较，但并未shijiyingy

尝试融合stride，但是有bug

修复bug
成功将不怕溢出的stride加入进代码

+++ os/src/config.rs
@@ -23,5 +23,5 @@ pub const TRAP_CONTEXT_BASE: usize = TRAMPOLINE - PAGE_SIZE;
 /// qemu board info
 pub use crate::board::{CLOCK_FREQ, MMIO};
 ///lab5 StrideInfo
-pub const INIT_PRIORITY:isize = 16;
-pub const BIGSTRDE:isize = isize::MAX >> 20; 
+pub const INIT_PRIORITY:u64 = 16;
+pub const BIGSTRDE:u64 = u64::MAX >> 20;
